thoughtleader-ai/requirements.txt
crewai
crewai-tools
python-dotenv
requests